Patch 1.0
UI scaling to just get the little dude to not overlap with the Logo
Moving the arrows from the tutorial onto the actual slide rather then below
Greying out the area of the background thats outside of the node grid. Currently you can draw beyond the grid which makes it confusing when the car stops following. Adding a black box with some opacity should make that process less confusing.
Spawn changes: Getting the obstacles and flags to spawn more consistently away from each other is huge I think. Most of the feedback is how the rocks spawned directly in the path of objectives, or flags spawned on objectives etc. Having a much wider distance between all of the spawning pieces would just make the game tighter.